Move main.cpp at the root of src
